proguard rules ㅊㅜ가

+++ bootpay/src/main/java/kr/co/bootpay/model/Item.kt
@@ -1,8 +1,8 @@
 package kr.co.bootpay.model
 
 class Item(
-        val name: String,
-        val quantity: Int,
-        val primaryKey: String,
-        val price: Double
+    val name: String,
+    val quantity: Int,
+    val primaryKey: String,
+    val price: Long
 )
